+++ src/main/java/insane96mcp/iguanatweaksreborn/network/message/StackSizesSync.java
@@ -12,14 +12,12 @@ import java.util.function.Supplier;
 public class StackSizesSync {
 	Boolean foodStackReduction;
 	String foodStackReductionFormula;
-	Integer stackableSoups;
 	Double itemStackMultiplier;
 	Double blockStackMultiplier;
 
-	public StackSizesSync(Boolean foodStackReduction, String foodStackReductionFormula, Integer stackableSoups, Double itemStackMultiplier, Double blockStackMultiplier) {
+	public StackSizesSync(Boolean foodStackReduction, String foodStackReductionFormula, Double itemStackMultiplier, Double blockStackMultiplier) {
 		this.foodStackReduction = foodStackReduction;
 		this.foodStackReductionFormula = foodStackReductionFormula;
-		this.stackableSoups = stackableSoups;
 		this.itemStackMultiplier = itemStackMultiplier;
 		this.blockStackMultiplier = blockStackMultiplier;
 	}
@@ -27,28 +25,26 @@ public class StackSizesSync {
 	public static void encode(StackSizesSync pkt, FriendlyByteBuf buf) {
 		buf.writeBoolean(pkt.foodStackReduction);
 		buf.writeUtf(pkt.foodStackReductionFormula);
-		buf.writeInt(pkt.stackableSoups);
 		buf.writeDouble(pkt.itemStackMultiplier);
 		buf.writeDouble(pkt.blockStackMultiplier);
 	}
 
 	public static StackSizesSync decode(FriendlyByteBuf buf) {
-		return new StackSizesSync(buf.readBoolean(), buf.readUtf(), buf.readInt(), buf.readDouble(), buf.readDouble());
+		return new StackSizesSync(buf.readBoolean(), buf.readUtf(), buf.readDouble(), buf.readDouble());
 	}
 
 	public static void handle(final StackSizesSync message, Supplier<NetworkEvent.Context> ctx) {
 		ctx.get().enqueueWork(() -> {
 			StackSizes.foodStackReduction = message.foodStackReduction;
 			StackSizes.foodStackReductionFormula = message.foodStackReductionFormula;
-			StackSizes.stackableSoups = message.stackableSoups;
 			StackSizes.itemStackMultiplier = message.itemStackMultiplier;
 			StackSizes.blockStackMultiplier = message.blockStackMultiplier;
 		});
 		ctx.get().setPacketHandled(true);
 	}
 
-	public static void sync(Boolean foodStackReduction, String foodStackReductionFormula, Integer stackableSoups, Double itemStackMultiplier, Double blockStackMultiplier, ServerPlayer player) {
-		Object msg = new StackSizesSync(foodStackReduction, foodStackReductionFormula, stackableSoups, itemStackMultiplier, blockStackMultiplier);
+	public static void sync(Boolean foodStackReduction, String foodStackReductionFormula, Double itemStackMultiplier, Double blockStackMultiplier, ServerPlayer player) {
+		Object msg = new StackSizesSync(foodStackReduction, foodStackReductionFormula, itemStackMultiplier, blockStackMultiplier);
 		NetworkHandler.CHANNEL.sendTo(msg, player.connection.connection, NetworkDirection.PLAY_TO_CLIENT);
 	}
 }